# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ActionApprovalPurpose": {
        "enum": [
          "EXECUTE",
          "ROLLBACK"
        ],
        "title": "ActionApprovalPurpose",
        "type": "string"
      },
      "ActionApprovalView": {
        "properties": {
          "action_request_id": {
            "format": "uuid",
            "title": "Action Request Id",
            "type": "string"
          },
          "approver_constraints": {
            "additionalProperties": true,
            "title": "Approver Constraints",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "plan_checksum_sha256": {
            "title": "Plan Checksum Sha256",
            "type": "string"
          },
          "purpose": {
            "$ref": "#/components/schemas/ActionApprovalPurpose"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requested_by": {
            "format": "uuid",
            "title": "Requested By",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ApprovalStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action_request_id",
          "purpose",
          "revision",
          "plan_checksum_sha256",
          "status",
          "reason",
          "requested_by",
          "approver_constraints",
          "decided_by",
          "decision_reason",
          "expires_at",
          "decided_at",
          "created_at",
          "updated_at"
        ],
        "title": "ActionApprovalView",
        "type": "object"
      },
...
      "ActionDetailView": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionRequestView"
          },
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/ActionApprovalView"
            },
            "title": "Approvals",
            "type": "array"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/ActionAttemptView"
            },
            "title": "Attempts",
            "type": "array"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionPlanView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action",
          "plan",
          "approvals",
          "attempts"
        ],
        "title": "ActionDetailView",
        "type": "object"
      },
...
      "ActionRequestView": {
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "cancellation_requested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Requested At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "plan_checksum_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Checksum Sha256"
          },
          "preflight": {
            "additionalProperties": true,
            "title": "Preflight",
            "type": "object"
          },
          "requested_by": {
            "format": "uuid",
            "title": "Requested By",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "rollback_parameters": {
            "additionalProperties": true,
            "title": "Rollback Parameters",
            "type": "object"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "target": {
            "additionalProperties": true,
            "title": "Target",
            "type": "object"
          },
          "timeout_seconds": {
            "title": "Timeout Seconds",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "action_type",
          "title",
          "description",
          "environment",
          "target",
          "parameters",
          "rollback_parameters",
          "status",
          "owner_id",
          "requested_by",
          "idempotency_key",
          "timeout_seconds",
          "deadline_at",
          "plan_checksum_sha256",
          "preflight",
          "result",
          "started_at",
          "completed_at",
          "cancellation_requested_at",
          "error_code",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "ActionRequestView",
        "type": "object"
      },
...
      "ApprovalStatus": {
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "EXPIRED",
          "CANCELLED"
        ],
        "title": "ApprovalStatus",
        "type": "string"
      },
...
      "Classification": {
        "enum": [
          "PUBLIC",
          "INTERNAL",
          "CONFIDENTIAL",
          "RESTRICTED"
        ],
        "title": "Classification",
        "type": "string"
      },
...
      "DecideActionApprovalRequest": {
        "properties": {
          "reason": {
            "maxLength": 4000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "DecideActionApprovalRequest",
        "type": "object"
      },
...
      "EventView": {
        "properties": {
          "causation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Causation Id"
          },
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sequence",
          "name",
          "run_id",
          "causation_id",
          "correlation_id",
          "schema_version",
          "classification",
          "payload",
          "created_at"
        ],
        "title": "EventView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreflightActionRequest": {
        "properties": {
          "approval_ttl_minutes": {
            "default": 60,
            "maximum": 1440.0,
            "minimum": 5.0,
            "title": "Approval Ttl Minutes",
            "type": "integer"
          },
          "reason": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "PreflightActionRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Enterprise Agent Runtime and Intelligence Workspace",
    "title": "Obsion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/action-approvals": {
      "get": {
        "operationId": "list_action_approvals_api_v1_action_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApprovalStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActionApprovalView"
                  },
                  "title": "Response List Action Approvals Api V1 Action Approvals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Action Approvals",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/action-approvals/{approval_id}/approve": {
      "post": {
        "operationId": "approve_action_api_v1_action_approvals__approval_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideActionApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionApprovalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/action-approvals/{approval_id}/reject": {
      "post": {
        "operationId": "reject_action_api_v1_action_approvals__approval_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideActionApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionApprovalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/{action_id}": {
      "get": {
        "operationId": "get_action_api_v1_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/{action_id}/cancel": {
      "post": {
        "operationId": "cancel_action_api_v1_actions__action_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionRequestView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/{action_id}/events": {
      "get": {
        "operationId": "list_action_events_api_v1_actions__action_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventView"
                  },
                  "title": "Response List Action Events Api V1 Actions  Action Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Action Events",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/{action_id}/preflight": {
      "post": {
        "operationId": "preflight_action_api_v1_actions__action_id__preflight_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreflightActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preflight Action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/{action_id}/rollback": {
      "post": {
        "operationId": "request_action_rollback_api_v1_actions__action_id__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionRequestView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Action Rollback",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/actions": {
      "get": {
        "operationId": "list_actions_api_v1_workspaces__workspace_id__actions_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ActionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActionRequestView"
                  },
                  "title": "Response List Actions Api V1 Workspaces  Workspace Id  Actions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Actions",
        "tags": [
          "actions"
        ]
      },
      "post": {
        "operationId": "create_action_api_v1_workspaces__workspace_id__actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionRequestView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Action",
        "tags": [
          "actions"
        ]
      }
    },
...
  }
}